tests/wpt/mozilla/tests/mozilla/structuredclone.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<iframe src="resources/iframe_contentDocument_inner.html" id="iframe"></iframe>
<script>
async_test(function(t) {
  var iframe = document.getElementById('iframe');
  var blob = new Blob(['test']);
  window.addEventListener("message", this.step_func_done(function(msg) {
      console.log('received msg')
      console.log(msg.data.size, blob.size)
      assert_equals(msg.data.size, blob.size);
      var otherWindow = iframe.contentWindow;
      otherWindow.postMessage(msg.data, '*');
  }), false);
  iframe.onload = this.step_func(function() {
      window.postMessage(blob, '*');
      console.log('sent msg')
  }, false);
}, "test cloning blobs");
</script>
</body>
</html>
